+++ CHANGELOG.md
@@ -44,6 +44,8 @@ The `Unreleased` section name is replaced by the expected version of next releas
 ### Removed
 ### Fixed
 
+- `Cosmos`: Fix defaulting for `compressUnfolds` in C# [#261](https://github.com/jet/equinox/pull/261)
+
 <a name="2.3.0"></a>
 ## [2.3.0] - 2020-11-04